Refatoração IA, e outros
Criada uma nova família de classes para lidar com o comportamento dos inimigos. Usando steering behaviour e states para controlar os ataques. Também criado um script de detecção do jogador. Agregado o sistema de Exp ao inimigo. Ativado os efeitos da skill tree para regen de vida e redução de dano

+++ Assets/Scripts/DataPersistence/Data/PlayerStatsData.cs
@@ -8,7 +8,7 @@ public class PlayerStatsData
     public int str;
     public int dex;
     public int inte;
-    public float exp;
+    public int exp;
     public int level;
     public float currentLife;
     public float baseLife;
@@ -60,7 +60,7 @@ public class PlayerStatsData
         this.baseLightAttackcDamage = playerStats.BaseLightAttackDamage;
         this.baseHeavyAttackDamage = playerStats.BaseHeavyAttackDamage;
     }
-    public PlayerStatsData(int con, int str,int dex, int inte, float exp, int level, float currentLife,float baseLife){
+    public PlayerStatsData(int con, int str,int dex, int inte, int exp, int level, float currentLife,float baseLife){
         this.con = con;
         this.str = str;
         this.dex = dex;
